refactor training and api architecture
